moved the copy_module_to_local functionality to HARK.core
